Database Schema
This database Schema consists of three tables (for now) that have names as User, Listings and Bookings. The User table has the various details of the user that is booking the Airbnb. The Listings table has details of the property listed on the platform for rent and the Bookings table displays various details regarding the user and the listing.